Move InkBar under tabs
