+++ src/tabs/tabs.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import TabTemplate from './tabTemplate';
-import InkBar from '../ink-bar';
+import InkBar from './InkBar';
 import getMuiTheme from '../styles/getMuiTheme';
 import warning from 'warning';
